Added user-prompted entry for Secret File generation.
Python's os.urandom *should* be an adequate entropy source, but
there is always the possibility that /dev/urandom on some system
will not be good, or that os.urandom itself will be changed to
use an insecure alternative on some system (for example, the
current os.urandom() on VMS calls OpenSSL RAND_pseudo_bytes which
is not documented as being a secure function.  However, looking at
the code it appears to call RAND_bytes - at least if there's no
engine configured - so I think it's OK but I'm not sure).  Bottom
line is there's enough layers of software between os.urandom and
good entropy sources that I feel better having an additional source.